mlops-local/requirements.txt
scikit-learn==1.6.1
pandas==2.2.3
joblib==1.4.2
mlflow==2.20.0
matplotlib
fastapi
uvicorn
boto3
python-dotenv